+++ src/lib/Main.h
@@ -59,11 +59,11 @@ class Main: public PdServ::Main {
 
         Parameter* addParameter( const char *path,
                 unsigned int mode, PdServ::Variable::Datatype datatype,
-                void *addr, size_t n, const unsigned int *dim);
+                void *addr, size_t n, const size_t *dim);
 
         Signal* addSignal( Task *task, unsigned int decimation,
                 const char *path, PdServ::Variable::Datatype datatype,
-                const void *addr, size_t n, const unsigned int *dim);
+                const void *addr, size_t n, const size_t *dim);
         int setParameter(const Parameter *p, size_t startIndex,
                 size_t nelem, const char *data,
                 struct timespec *) const;
